Added db fix
Added db fix feature:

In math and physics, 0 input is not possible when db measurement is taken.

This option will enable a simple fix for songs/tracks 0 samples by replacing
each sample with input 0 with input of 1 bit which is 1.

This make the song/track feels alive also correct in math/physics.

This will replace 0 sample into 1 bit sample, so it will be ~ 0,00305 % noise of amptitude
instead of 0. AND: 0 DB instead of infinity db. This is correct.